% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{ComputeLinkagedisequilibrium}
\alias{ComputeLinkagedisequilibrium}
\title{Title Compute Linkage disequilibrium between two snps or two locus}
\usage{
ComputeLinkagedisequilibrium(
  data,
  Snp1,
  Snp2,
  LinkageDesiquilibriumType = "D",
  verbose = FALSE
)
}
\arguments{
\item{data}{matrix
ncols at least 2 cols
n row = N individuals x 2 haplotypes}

\item{Snp1}{SNP present int the data matrix}

\item{Snp2}{SNP present int the data matrix}

\item{LinkageDesiquilibriumType}{type of method use to compute linkage disequilibrium
value in D, Dprime, Rsquare, I, NI, NI2, NI3
D
D prime (Lewotin RC 1964)
R square (Hill et Robertson 1968)
I Shannon Entropy
NI normalized mutual information}

\item{verbose}{print or not computation value False by default}
}
\value{
single value of linkage disequilibrium
}
\description{
Title Compute Linkage disequilibrium between two snps or two locus
}
\examples{

Example<-structure(
c("G", "G", "A", "G", "G",
  "A", "G", "G", "A", "G",
  "C", "C", "C", "C", "C",
  "T", "C", "C", "C", "C",
  "G", "A", "A", "A", "A"),
  .Dim = c(5L, 5L),
  .Dimnames = list(
  c("Ind1_A","Ind1_B",
   "Ind2_A", "Ind2_B",
   "Ind3_A"),
   c("rs3", "rs1", "rs4", "rs44", "rs7" )))
ComputeLinkagedisequilibrium(Example,"rs3","rs44",LinkageDesiquilibriumType="Rsquare")
}
