% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequencePlay.R
\name{EnumerateAllHapsOfN}
\alias{EnumerateAllHapsOfN}
\title{Title}
\usage{
EnumerateAllHapsOfN(VCFStructure, NSnps = 2)
}
\arguments{
\item{VCFStructure}{The structure obtained for this packages}

\item{NSnps}{Number of SNPS to take account for this list}
}
\value{
data frame for of the different combination of NSnps and the max
difference for all observed
}
\description{
Title
}
\examples{
library("HapToolkit")
VCF_file <- system.file("extdata",
 "TEST.vcf",
 package = "HapToolkit")
 FAM_file <- system.file("extdata",
 "TEST.fam",
 package = "HapToolkit")
I<-ReadPhasedVCF(VCF_file,FAM_file)
HAPS<-EnumerateAllHapsOfN(I,2);HAPS[order(HAPS$MaxDiff,decreasing = TRUE),]
start_time <- Sys.time()
HAP<-data.frame()
for (i in 2:7){
   cat(i,"\n")
  TEMP<-EnumerateAllHapsOfN(I,i)
  HAP<-rbind(HAP,TEMP)
}
end_time <- Sys.time()
end_time - start_time
# start_time <- Sys.time()
# HAP<-data.frame()
# ZeMaxDiff<-0
# for (i in 2:length(I$SNP_LIST)){
#   cat(i,"\n")
#   TEMP<-EnumerateAllHapsOfN(I,i)
#   if(max(TEMP$MaxDiff)<ZeMaxDiff){
#     break
#   }else{
#     ZeMaxDiff<-max(TEMP$MaxDiff)
#     HAP<-rbind(HAP,TEMP)
#   }
# }
# end_time <- Sys.time()
# end_time - start_time
}
