% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HaploBlock.R
\name{MakeSNPBlockTree}
\alias{MakeSNPBlockTree}
\title{Make SNP Block Tree make a tree of haplotype block based on normalized mutual
information}
\usage{
MakeSNPBlockTree(HAP)
}
\arguments{
\item{HAP}{haplotypye matrix 2 N ind rows x M snps columns}
}
\description{
Make SNP Block Tree make a tree of haplotype block based on normalized mutual
information
}
\examples{
library(HapToolkit)
HAP<-structure(list(
rs1 = c("C", "C", "C", "C", "T", "C", "C", "C", "C", "C", "C", "C"),
rs4 = c("G", "G", "G", "G", "A", "G", "G", "G", "G", "G", "G", "G"),
rs12 = c("G", "G", "G", "G", "A", "G", "G", "G", "G", "G", "G", "G"),
rs7 = c("A", "A", "A", "A", "C", "A", "A", "A", "A", "A", "A", "A"),
rs14 = c("G", "A", "A", "G", "A", "G", "A", "G", "A", "G", "G", "A"),
rs75 = c("C", "T", "T", "C", "T", "C", "T", "C", "T", "C", "C", "T"),
rs13 = c("C", "T", "T", "C", "T", "C", "T", "C", "T", "C", "C", "T"),
rs3 = c("A", "A", "A", "A", "A", "A", "G", "A", "A", "A", "A", "A"),
rs19 = c("T", "T", "T", "T", "T", "T", "C", "T", "T", "T", "T", "T"),
rs21 = c("T", "C", "C", "T", "T", "T", "T", "C", "C", "T", "C", "C")),
row.names = c("Ind1_A", "Ind1_B", "Ind2_A", "Ind2_B",
 "Ind3_A", "Ind3_B", "Ind4_A", "Ind4_B", "Ind5_A",
  "Ind5_B", "Ind6_A", "Ind6_B"), class = "data.frame")

HapTree<-MakeSNPBlockTree(HAP)
print(HapTree$DataTree)
library(igraph)
#library(networkD3)
#plot(as.igraph(HapTree$DataTree,directed=TRUE,direction="climb"))
#HapTreeNetwork <- ToDataFrameNetwork(HapTree$DataTree, "name")
#simpleNetwork(HapTreeNetwork[-3], fontSize = 12)
}
