% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{PlotLinkagedisequilibriumMatrix}
\alias{PlotLinkagedisequilibriumMatrix}
\title{Title plot a Linkage desiquilibrium matrix as a ggplot2 graph}
\usage{
PlotLinkagedisequilibriumMatrix(LD_MATRIX, LinkageDesiquilibriumType)
}
\arguments{
\item{LD_MATRIX}{matrix of linkage disequilibrium, suppose to be a square matrix}

\item{LinkageDesiquilibriumType}{type of method use to compute linkage disequilibrium
value in D, Dprime, Rsquare, I, NI, NI2, NI3
D
D prime (Lewotin RC 1964)
R square (Hill et Robertson 1968)
I Shannon Entropy
NI normalized mutual information}
}
\value{
G a ggplot2 graph of the linkage desiquilibrium
}
\description{
Title plot a Linkage desiquilibrium matrix as a ggplot2 graph
}
\examples{
library(HapToolkit)
library(plotly)
HAP<-structure(list(
rs1 = c("C", "C", "C", "C", "T", "C", "C", "C", "C", "C", "C", "C"),
rs4 = c("G", "G", "G", "G", "A", "G", "G", "G", "G", "G", "G", "G"),
rs12 = c("G", "G", "G", "G", "A", "G", "G", "G", "G", "G", "G", "G"),
rs7 = c("A", "A", "A", "A", "C", "A", "A", "A", "A", "A", "A", "A"),
rs14 = c("G", "A", "A", "G", "A", "G", "A", "G", "A", "G", "G", "A"),
rs75 = c("C", "T", "T", "C", "T", "C", "T", "C", "T", "C", "C", "T"),
rs13 = c("C", "T", "T", "C", "T", "C", "T", "C", "T", "C", "C", "T"),
rs3 = c("A", "A", "A", "A", "A", "A", "G", "A", "A", "A", "A", "A"),
rs19 = c("T", "T", "T", "T", "T", "T", "C", "T", "T", "T", "T", "T"),
rs21 = c("T", "C", "C", "T", "T", "T", "T", "C", "C", "T", "C", "C")),
row.names = c("Ind1_A", "Ind1_B", "Ind2_A", "Ind2_B",
 "Ind3_A", "Ind3_B", "Ind4_A", "Ind4_B", "Ind5_A",
  "Ind5_B", "Ind6_A", "Ind6_B"), class = "data.frame")

LD<-ComputeLinkagedisequilibriumMatrix(HAP,"Rsquare")
ggplotly(PlotLinkagedisequilibriumMatrix(LD))
}
