% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{ReadPhasedVCF}
\alias{ReadPhasedVCF}
\title{ReadPhasedVCF}
\usage{
ReadPhasedVCF(filename, famfile = "", chr = 0, start = 0, end = 0)
}
\arguments{
\item{filename}{name of a phased vcf as for example
##fileformat=VCFv4.2
#FILTER=<ID=PASS,Description="All filters passed">
#fileDate=24/05/2023 - 12:10:24
#source=shapeit4.1.3
#contig=<ID=1>
#INFO=<ID=AF,Number=A,Type=Float,Description="Allele Frequency">
#INFO=<ID=AC,Number=1,Type=Integer,Description="Allele count">
#INFO=<ID=CM,Number=A,Type=Float,Description="Interpolated cM position">
#FORMAT=<ID=GT,Number=1,Type=String,Description="Phased genotypes">
#CHROM  POS     ID      REF     ALT     QUAL    FILTER  INFO    FORMAT  IndA       IndB   IndC
1       944564  rs3128117       T       C       .       .       AC=269;AF=0.340506;CM=0.191843  GT      1|0     0|0     1|1}

\item{famfile}{.fam file used for Shape IT, as used in plink}

\item{chr}{chromosome of interest}

\item{start}{begening of the block of interest (position included)}

\item{end}{end of the block of interest}
}
\value{
list of data table
}
\description{
Read a VCF and generate an haplotype matrix
}
\examples{
library("HapToolkit")
VCF_file <- system.file("extdata",
 "TEST.vcf",
 package = "HapToolkit")
 FAM_file <- system.file("extdata",
 "TEST.fam",
 package = "HapToolkit")#'
I<-ReadPhasedVCF(VCF_file)
B<-ReadPhasedVCF(VCF_file,FAM_file)
C<-ReadPhasedVCF(VCF_file,FAM_file,chr=1,start=838555,end=891945)
filename<-VCF_file
famfile<-FAM_file
}
